--- a/app/views/contabilidad/reportes/_account.html.slim
+++ b/app/views/contabilidad/reportes/_account.html.slim
@@ -13,13 +13,12 @@ table
 
       - accounts.each do |account|
         - balance = account.balance(from_date: @from_date, to_date: @to_date)
-        - if balance != 0
-          tr
-            td align="center"
-            td=account.name
-            td=balance.round(2)
-            - running_total += balance
-            td=running_total.round(2)
+        tr
+          td align="center"
+          td=account.name
+          td=balance.round(2)
+          - running_total += balance
+          td=running_total.round(2)
 
       tr
         td align="center"