--- a/app/views/contabilidad/reportes/_cash_flow_accounts.html.slim
+++ b/app/views/contabilidad/reportes/_cash_flow_accounts.html.slim
@@ -6,9 +6,9 @@ tr
   tr
     td
       em=account[:name]
-    - account[:periods].each do |period|
+    - account[:periods].each do |month, balance|
       td
-        em=link_to_account( period[:balance], data[:epoch] ? (period[:month] - 1.day).beginning_of_month : period[:month], account[:id] )
+        em=link_to_account( balance, data[:epoch] ? (month - 1.day).beginning_of_month : month, account[:id] )
 tr
   th.text-right scope="row" Total de #{name}
   - data[:totals].each do |balance|