--- a/config/initializers/plutus.rb
+++ b/config/initializers/plutus.rb
@@ -5,7 +5,10 @@ require 'extensions/plutus'
 Rails.application.config.to_prepare do
   begin
     if ActiveRecord::Base.connection.table_exists?(:plutus_entries)
-      Plutus::Entry.include Extensions::Plutus
+      Plutus::Entry.include Extensions::PlutusEntry
+      Plutus::DebitAmount.include Extensions::PlutusAmount
+      Plutus::CreditAmount.include Extensions::PlutusAmount
+      Plutus::Account.include Extensions::PlutusAccount
     end
   rescue ActiveRecord::NoDatabaseError
   end