--- a/lib/extensions/plutus.rb
+++ b/lib/extensions/plutus.rb
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 module Extensions
-  module Plutus
+  module PlutusEntry
     extend ActiveSupport::Concern
     
     included do
@@ -15,6 +15,58 @@ module Extensions
       scope :with_description, ->(description) {
         where("description like ?", "%#{description}%") if description.present?
       }
+    end
+  end
+
+  module PlutusAmount
+    extend ActiveSupport::Concern
+    
+    class_methods do
+      def balance_by_month(year)
+        includes(:entry).group_by_month(
+          :date,
+          range: year.beginning_of_year..year.end_of_year,
+          time_zone: "UTC"
+        ).sum(:amount)
+      end
+    end
+  end
+
+  module PlutusAccount
+    extend ActiveSupport::Concern
+
+    def balance_by_month(year)
+      if self.class == Plutus::Account
+        raise(NoMethodError, "undefined method 'balance_by_month'")
+      else
+        if self.normal_credit_balance ^ contra
+          debit_amounts.balance_by_month(year).merge(credit_amounts.balance_by_month(year)) { |_, a, b| b - a } 
+        else
+          debit_amounts.balance_by_month(year).merge(credit_amounts.balance_by_month(year)) { |_, a, b| a - b } 
+        end
+      end
+    end
+
+    class_methods do
+      def balance_by_month(year)
+        if self.new.class == Plutus::Account
+          raise(NoMethodError, "undefined method 'balance_by_month'")
+        else
+          accounts_balance = {}
+          accounts = self.all
+          accounts.each do |account|
+            if account.contra
+              puts "Contra"
+              puts account.balance_by_month(year)
+              accounts_balance.merge!(account.balance_by_month(year)) { |_, a, b| a - b }
+            else
+              puts account.balance_by_month(year)
+              accounts_balance.merge!(account.balance_by_month(year)) { |_, a, b| a + b }
+            end
+          end
+          accounts_balance
+        end
+      end
     end
   end
 end