Completed 200 OK in 3947ms (Views: 3351.3ms | ActiveRecord: 528.1ms | Allocations: 2443810)

Completed 200 OK in 1564ms (Views: 50.6ms | ActiveRecord: 468.5ms | Allocations: 999253)

Completed 200 OK in 282ms (Views: 42.6ms | ActiveRecord: 75.1ms | Allocations: 193241)